- Reviewing and fixing scrollable layout solution - Rewriting existing pages of Signature request and connect account to match Figma Design
