+++ src/apps/connect-to-app/index.html
@@ -11,13 +11,8 @@
   />
 
   <title>Casper Wallet: Connect to Site</title>
-  <style>
-    div#connect-to-app-app-container {
-        height: 100%;
-    }
-  </style>
 </head>
 <body>
-  <div id="connect-to-app-app-container"></div>
+  <div id="app-container"></div>
 </body>
 </html>
