+++ src/apps/connect-to-app/pages/accounts-selection/index.tsx
@@ -1,164 +1,71 @@
-import React, { useMemo, useCallback, SetStateAction, Dispatch } from 'react';
-import styled from 'styled-components';
+import React, { Dispatch, SetStateAction } from 'react';
+import { Trans, useTranslation } from 'react-i18next';
 
 import {
-  Button,
-  Checkbox,
-  SiteFaviconBadge,
-  Hash,
-  HashVariant,
-  List,
-  Typography
-} from '@libs/ui';
-
-import {
-  PageContainer,
-  ContentContainer,
-  HeaderTextContainer,
-  ListItemClickableContainer,
   FooterButtonsContainer,
-  LeftAlignedFlexColumn
-} from '@src/libs/layout/containers';
+  HeaderSubmenuBarNavLink,
+  LayoutWindow,
+  PopupHeader
+} from '@src/libs/layout';
+import { Button, Typography } from '@src/libs/ui';
 
-import { useTranslation, Trans } from 'react-i18next';
-import { useSelector } from 'react-redux';
-import {
-  selectVaultAccounts,
-  selectVaultActiveAccount
-} from '@src/background/redux/vault/selectors';
-
-import { RouterPath, useTypedNavigate } from '@connect-to-app/router';
-
-const HeaderTextContent = styled.div`
-  margin-top: 16px;
-`;
+import { RouterPath, useTypedNavigate } from '../../router';
+import { AccountsSelectionContent } from './content';
+import styled from 'styled-components';
 
 const TextCentredContainer = styled.div`
   text-align: center;
 `;
 
-const AccountBalanceListItemContainer = styled(LeftAlignedFlexColumn)``;
-const AccountNameWithHashListItemContainer = styled(LeftAlignedFlexColumn)`
-  width: 100%;
-`;
-
-interface AccountsSelectionPageProps {
+export interface Props {
   selectedAccountNames: string[];
   setSelectedAccountNames: Dispatch<SetStateAction<string[]>>;
   origin: string;
-  headerText: string;
+  title: string;
 }
 
 export function AccountsSelectionPage({
   selectedAccountNames,
   setSelectedAccountNames,
   origin,
-  headerText
-}: AccountsSelectionPageProps) {
-  const navigate = useTypedNavigate();
+  title
+}: Props) {
   const { t } = useTranslation();
-
-  const accounts = useSelector(selectVaultAccounts);
-  const activeAccount = useSelector(selectVaultActiveAccount);
-
-  const handleSelectAll = useCallback(() => {
-    setSelectedAccountNames(accounts.map(account => account.name));
-  }, [accounts, setSelectedAccountNames]);
-
-  const handleUnselectAll = useCallback(() => {
-    setSelectedAccountNames([]);
-  }, [setSelectedAccountNames]);
-
-  const areAllAccountsSelected = accounts.every(account =>
-    selectedAccountNames.includes(account.name)
-  );
-
-  const headerAction = useMemo(() => {
-    const captionSelectAll = t('select all');
-    const captionUnselectAll = t('unselect all');
-
-    return areAllAccountsSelected
-      ? { caption: captionUnselectAll, onClick: handleUnselectAll }
-      : { caption: captionSelectAll, onClick: handleSelectAll };
-  }, [t, areAllAccountsSelected, handleSelectAll, handleUnselectAll]);
-
-  const accountsListItems = accounts.map(account => ({
-    ...account,
-    id: account.name
-  }));
+  const navigate = useTypedNavigate();
 
   return (
-    <PageContainer>
-      <ContentContainer>
-        <HeaderTextContainer>
-          <SiteFaviconBadge origin={origin} />
-          <HeaderTextContent>
-            <Typography type="header">
-              <Trans t={t}>{headerText}</Trans>
-            </Typography>
-          </HeaderTextContent>
-        </HeaderTextContainer>
-        <List
-          headerLabel={t('select account(s)')}
-          headerAction={headerAction}
-          rows={accountsListItems}
-          renderRow={account => (
-            <ListItemClickableContainer
-              onClick={() =>
-                setSelectedAccountNames(selectedAccountNames =>
-                  selectedAccountNames.includes(account.name)
-                    ? selectedAccountNames.filter(
-                        accountName => accountName !== account.name
-                      )
-                    : [...selectedAccountNames, account.name]
-                )
-              }
-            >
-              <Checkbox
-                variant="square"
-                checked={selectedAccountNames.includes(account.name)}
-              />
-              <AccountNameWithHashListItemContainer>
-                <Typography
-                  type={
-                    activeAccount && activeAccount.name === account.name
-                      ? 'bodySemiBold'
-                      : 'body'
-                  }
-                >
-                  {account.name}
-                </Typography>
-                <Hash
-                  value={account.publicKey}
-                  variant={HashVariant.CaptionHash}
-                  truncated
-                />
-              </AccountNameWithHashListItemContainer>
-
-              <AccountBalanceListItemContainer>
-                <Typography type="bodyHash">2.1M</Typography>
-                <Typography type="bodyHash" color="contentSecondary">
-                  CSPR
-                </Typography>
-              </AccountBalanceListItemContainer>
-            </ListItemClickableContainer>
+    <LayoutWindow
+      Header={
+        <PopupHeader
+          withConnectionStatus
+          renderSubmenuBarItems={() => (
+            <HeaderSubmenuBarNavLink linkType="cancelWindow" />
           )}
-          marginLeftForItemSeparatorLine={60}
         />
-      </ContentContainer>
-      <FooterButtonsContainer>
-        <TextCentredContainer>
-          <Typography type="captionRegular">
-            <Trans t={t}>Only connect with sites you trust</Trans>
-          </Typography>
-        </TextCentredContainer>
-        <Button
-          disabled={selectedAccountNames.length === 0}
-          onClick={() => navigate(RouterPath.ApproveConnection)}
-        >
-          <Trans t={t}>Next</Trans>
-        </Button>
-      </FooterButtonsContainer>
-    </PageContainer>
+      }
+      Content={
+        <AccountsSelectionContent
+          selectedAccountNames={selectedAccountNames}
+          setSelectedAccountNames={setSelectedAccountNames}
+          origin={origin}
+          headerText={title}
+        />
+      }
+      renderFooter={() => (
+        <FooterButtonsContainer>
+          <TextCentredContainer>
+            <Typography type="captionRegular">
+              <Trans t={t}>Only connect with sites you trust</Trans>
+            </Typography>
+          </TextCentredContainer>
+          <Button
+            disabled={selectedAccountNames.length === 0}
+            onClick={() => navigate(RouterPath.ApproveConnection)}
+          >
+            <Trans t={t}>Next</Trans>
+          </Button>
+        </FooterButtonsContainer>
+      )}
+    />
   );
 }
