+++ src/apps/import-account-with-file/index.html
@@ -10,9 +10,9 @@
     rel="stylesheet"
   />
 
-  <title>Casper Wallet: Import account with file</title>
+  <title>Casper Wallet: Import account</title>
 </head>
 <body>
-  <div id="import-account-with-file-app-container"></div>
+  <div id="app-container"></div>
 </body>
 </html>
