+++ src/apps/signature-request/app.tsx
@@ -1,7 +1,5 @@
 import React from 'react';
 
-import { LayoutWindow, PopupHeader } from '@libs/layout';
-
 import { SignatureRequestPage } from './pages/signature-request';
 import { HashRouter, Route, Routes } from 'react-router-dom';
 
@@ -9,15 +7,7 @@ export function App() {
   return (
     <HashRouter>
       <Routes>
-        <Route
-          path="*"
-          element={
-            <LayoutWindow
-              Header={<PopupHeader withConnectionStatus />}
-              Content={<SignatureRequestPage />}
-            />
-          }
-        />
+        <Route path="*" element={<SignatureRequestPage />} />
       </Routes>
     </HashRouter>
   );
