+++ src/apps/signature-request/index.html
@@ -13,6 +13,6 @@
   <title>Casper Wallet: Signature Request</title>
 </head>
 <body>
-  <div id="signature-request-app-container"></div>
+  <div id="app-container"></div>
 </body>
 </html>
