+++ src/libs/ui/global-style.ts
@@ -10,10 +10,11 @@ export const GlobalStyle = createGlobalStyle<{ theme: any }>`
   
   body {
     height: 100%;
-    
-    background: none;
+    min-height: 600px;
+    min-width: 360px;
     padding: 0;
-    
+
+    background: none;
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
   }
@@ -23,6 +24,7 @@ export const GlobalStyle = createGlobalStyle<{ theme: any }>`
   }
   
   #app-container {
+    height: 100%;
     background: ${props => props.theme.color.backgroundSecondary};
     border-radius: 16px;
   }
